% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stack_dirich.R
\name{data_stack_dirich}
\alias{data_stack_dirich}
\title{Preparing the data}
\usage{
data_stack_dirich(y, covariates, share = NULL, data, d, n)
}
\arguments{
\item{y}{Response variable in a matrix format.}

\item{covariates}{String with the name of covariates.}

\item{share}{Covariates to share in all the cateogries. TODO}

\item{data}{Data.frame which contains all the covariates.}

\item{d}{Number of categories.}

\item{n}{Number of locations.}
}
\value{
Matrix A such as eta = A \%*\%x
}
\description{
\code{data_stack_dirich} prepares the data using inla.stack from the package INLA.
}
\examples{
n <- 100
d <- 4

V <- matrix(rnorm(4*n, 0, 1), ncol=4)
V <- as.data.frame(V)
names(V) <- c('v1', 'v2', 'v3', 'v4' )

covariates <- names(V)


formula <- y ~ 1 + v1 + v2 | 1 + v1 | 1 + v1

names_cat <- formula_list(formula)

data_stack_construct <- data_stack_dirich(y          = as.vector(rep(NA, n*d)),
                                          covariates = names_cat,
                                          share      = NULL,
                                          data       = V,
                                          d          = d,
                                          n          = n )

}
\author{
Joaquín Martínez-Minaya \email{jomarminaya@gmail.com}
}
