% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirinla-package.R
\docType{package}
\name{dirinla-package}
\alias{dirinla-package}
\alias{dirinla}
\title{dirinla}
\description{
The R-package dirinla allows the user to fit models in the compositional data
context. In particular, it allows fit Dirichlet regression models using the
Integrated Nested Laplace Approximation (INLA) methodology.
}
\details{
See \code{\link[=dirinlareg]{dirinlareg()}}.

Tutorials and more information can be found at
\url{https://inlabru-org.github.io/dirinla/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://inlabru-org.github.io/dirinla/}
  \item Report bugs at \url{https://github.com/inlabru-org/dirinla/issues}
}

}
\author{
Joaquín Martínez-Minaya \email{jomarminaya@gmail.com}
and Finn Lindgren \email{finn.lindgren@gmail.com}
}
