% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_data}
\alias{draw_data}
\title{Draw samples from the prior distribution}
\usage{
draw_data(x, d, V)
}
\arguments{
\item{n:}{the number of simulations.}

\item{x:}{parameters of the latent field.}

\item{d:}{the dimension of the response variable}

\item{V:}{matrix with covariates corresponding with the parameters}
}
\value{
data.frame with the simulations
}
\description{
`draw_data` Function to draw samples for a multivariate
normal prior distribution.
}
\author{
Joaquín Martínez-Minaya <\email{joaquin.martinez-minaya@uv.es}>
}
