% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_prior.R
\name{draw_prior}
\alias{draw_prior}
\title{Draw samples from the prior distribution}
\usage{
draw_prior(n, Q)
}
\arguments{
\item{n:}{Number of samples.}

\item{Q:}{Precision matrix of the multivariate normal distribution.}

\item{...:}{Arguments of the mvrnorm function.}
}
\value{
Data.frame with the simulations.
}
\description{
`draw_prior` draws samples from a multivariate normal prior distribution.
}
\author{
Joaquín Martínez-Minaya <\email{joaquin.martinez-minaya@uv.es}>
}
