% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g0_vector_eta.R
\name{g0_vector_eta}
\alias{g0_vector_eta}
\title{Computing gradient vector in eta}
\usage{
g0_vector_eta(A = A, x, y)
}
\arguments{
\item{A}{Matrix which links eta with the latent field, i.e., eta = A x.}

\item{x}{Vector with the elements of the latent field, i.e., eta = A x.}

\item{y}{Vector with the response variable.}
}
\value{
A numeric vector with the gradient in eta.
}
\description{
`g0_vector_eta` computes the gradient of -loglikelihood
}
\author{
Joaquín Martínez-Minaya <\email{joaquin.martinez-minaya@uv.es}>
}
