% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_matrix_eta.R
\name{H_matrix_eta}
\alias{H_matrix_eta}
\title{Computing the real Hessian in eta}
\usage{
H_matrix_eta(A = A, x, y)
}
\arguments{
\item{A:}{Matrix which links eta with the latent field, i.e., eta = A x.}

\item{x:}{Vector with the elements of the latent field, i.e., eta = A x.}

\item{y:}{Vector with dimension 1 x d.}
}
\value{
Expected Hessian in eta.
}
\description{
`H_matrix_eta` computes the real Hessian in eta of -loglikelihood
}
\author{
Joaquín Martínez-Minaya <\email{joaquin.martinez-minaya@uv.es}>
}
