% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiag_m.R
\name{bdiag_m}
\alias{bdiag_m}
\title{Block diagonal matrix creation}
\usage{
bdiag_m(lmat)
}
\arguments{
\item{lmat}{\verb{list(<mat1>, <mat2>, ....., <mat_N>)}  where each \code{mat_j} is a  \verb{k x k} 'matrix'}
}
\value{
a sparse (N\emph{k x N}k) matrix of class \link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix}.
}
\description{
Fast version of \code{\link[Matrix:.bdiag]{Matrix::.bdiag()}} -- for the case of \emph{many}  (k x k) matrices:
Copyright (C) 2016 Martin Maechler, ETH Zurich
}
