% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiag_m.R
\name{bdiag_m}
\alias{bdiag_m}
\title{Fast version of Matrix :: .bdiag() -- for the case of *many*  (k x k) matrices:
Copyright (C) 2016 Martin Maechler, ETH Zurich}
\usage{
bdiag_m(lmat)
}
\arguments{
\item{lmat}{list(<mat1>, <mat2>, ....., <mat_N>)  where each mat_j is a  k x k 'matrix'}
}
\value{
a sparse (N*k x N*k) matrix of class  \code{"\linkS4class{dgCMatrix}"}.
}
\description{
Fast version of Matrix :: .bdiag() -- for the case of *many*  (k x k) matrices:
Copyright (C) 2016 Martin Maechler, ETH Zurich
}
