% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digamma.R
\name{digamma_red}
\alias{digamma_red}
\title{Computing the function digamma}
\usage{
digamma_red(x, ...)
}
\arguments{
\item{x}{Argument to applied the function digamma.}

\item{...}{Rest of arguments used in the case of digamma functions.}
}
\value{
Result of applying digamma function
}
\description{
\code{digamma_red} is the function digamma appropiate for really small values
}
\author{
Joaquín Martínez-Minaya \email{joaquin.martinez-minaya@uv.es}
}
