% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_log_pos_x.R
\name{dirichlet_log_pos_x}
\alias{dirichlet_log_pos_x}
\title{Dirichlet log posterior function}
\usage{
dirichlet_log_pos_x(A = A, x, Qx = Qx, y)
}
\arguments{
\item{A}{A matrix which links eta with the latent field, i.e., eta = A x.}

\item{x}{Vector with the elements of the latent field, i.e., eta = A x.}

\item{Qx}{Precision matrix for the priors of the latent field.}

\item{y}{Vector with the response variable.}
}
\value{
A real value showing the -log posterior density is returned
}
\description{
`dirichlet_log_pos_x` returns the -log posterior Dirichlet distribution asumming
multivariate normal prior with precision matrix Qx for elements of the latent field.
}
\author{
Joaquín Martínez-Minaya <\email{joaquin.martinez-minaya@uv.es}>
}
