% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirinlareg.R
\name{dirinlareg}
\alias{dirinlareg}
\title{Fitting a Dirichlet regression}
\usage{
dirinlareg(
  formula,
  y,
  data.cov,
  share = NULL,
  x0 = NULL,
  tol0 = 1e-05,
  tol1 = 0.1,
  k0 = 20,
  k1 = 5,
  a = 0.5,
  strategy = "ls-quasi-newton",
  prec = prec,
  verbose = FALSE,
  cores = 1,
  sim = 1000,
  prediction = FALSE,
  data.pred.cov = NULL,
  ...
)
}
\arguments{
\item{formula}{object of class formula indicating the response variable and the covariates of the Dirichlet regression}

\item{y}{matrix containing the response variable \eqn{R^{nxd}}, being n number of individuals
and d the number of categories}

\item{data.cov}{data.frame with the covarites, only the covariates!}

\item{share}{parameters to be fitted jointly.}

\item{x0}{initial optimization value}

\item{tol0}{tolerance}

\item{tol1}{tolerance for the gradient such that |grad| < tol1 * max(1, |f|)}

\item{k0}{number of iterations}

\item{k1}{number of iterations including the calling to inla}

\item{a}{step length in the optimization algorithm}

\item{strategy}{strategy to use to optimize}

\item{prec}{precision for the prior of the fixed effects}

\item{verbose}{if TRUE all the computing process is shown. Default is FALSE}

\item{cores}{Number of cores for parallel computation. The package parallel is used.}

\item{sim}{Simulations to call inla.posterior.sample and extract linear predictor, alphas and mus.
The bigger it is, better is the approximation, but more computational time.}

\item{prediction}{if TRUE we will predict with the new values of the covariates given in data.pred.cov.}

\item{data.pred.cov}{data.frame with the values for the covariates where we want to predict.}

\item{...}{arguments for the inla command}
}
\value{
model dirinlaregmodel object
}
\description{
\code{dirinlareg} Main function to do a Dirichlet Regression
}
\examples{
if (dirinla_safe_inla() &&
    requireNamespace("DirichletReg", quietly = TRUE)) {
### In this example, we show how to fit a model using the dirinla package ###
### --- 1. Loading the libraries --- ####


### --- 2. Simulating from a Dirichlet likelihood --- ####
set.seed(1000)
N <- 50 #number of data
V <- as.data.frame(matrix(runif((4) * N, 0, 1), ncol = 4)) #Covariates
names(V) <- paste0('v', 1:4)

formula <- y ~ 1 + v1 | 1 + v2 | 1 + v3 | 1 + v4
(names_cat <- formula_list(formula))

x <- c(-1.5, 1, -3, 1.5,
       2, -3 , -1, 5)

mus <- exp(x) / sum(exp(x))
C <- length(names_cat)
data_stack_construct <-
  data_stack_dirich(y = as.vector(rep(NA, N * C)),
                    covariates = names_cat,
                    data       = V,
                    d          = C,
                    n          = N)

A_construct <- data_stack_construct
A_construct[1:8, ]

eta <- A_construct \%*\% x
alpha <- exp(eta)
alpha <- matrix(alpha,
                ncol  = C,
                byrow = TRUE)
y_o <- DirichletReg::rdirichlet(N, alpha)
colnames(y_o) <- paste0("y", 1:C)
head(y_o)


### --- 3. Fitting the model --- ####
y <- y_o
model.inla <- dirinlareg(
  formula  = y ~ 1 + v1 | 1 + v2 | 1 + v3 | 1 + v4,
  y        = y,
  data.cov = V,
  prec     = 0.0001,
  verbose  = FALSE)


summary(model.inla)
}
}
\author{
Joaquín Martínez-Minaya \email{jomarminaya@gmail.com}
}
