% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirinlareg.R
\name{dirinlareg}
\alias{dirinlareg}
\title{Fitting a Dirichlet regression}
\usage{
dirinlareg(formula, y, data.cov, share = NULL, x0 = NULL,
  tol0 = 1e-10, tol1 = 1e-10, k0 = 100, a = 0.5,
  strategy = "ls-quasi-newton", prec = prec, verbose = FALSE, ...)
}
\arguments{
\item{y:}{matrix containing the response variable R^{nxd}, being n number of individuals
and d the number of categories}

\item{data.cov:}{data.frame with the covarites, only the covariates!}

\item{share:}{parameters to be fitted jointly.}

\item{x0:}{initial optimization value}

\item{tol0:}{tolerance}

\item{tol1:}{tolerance for the gradient such that |grad| < tol1 * max(1, |f|)}

\item{k0:}{number of iterations}

\item{a:}{step length in the optimization algorithm}

\item{strategy:}{strategy to use to optimize}

\item{Qx:}{matrix of precision for the prior Gaussian field}

\item{...:}{}
}
\value{
model: inla object

mean : posteriors means of the parameters corresponding to the latent variables
}
\description{
`dirinlareg` Main function to do a Dirichlet Regression
}
\examples{
### --- 1. Data simulation --- ####
set.seed(1000)
n <- 100
x <- c(-2, -2, -2,  3, + 3, + 3, 2,2,2)

## Latent field
d <- 3
m <- length(x)/d #including intercept

## Generate covariates
v <- rnorm(n, 0,1)
V <- as.data.frame(matrix(rnorm((m-1)*n, 0, 1), ncol=m-1))
names(V) <- paste0('v', 1:(m-1))

## Response variable
y_o <- draw_data(x = x,
                 d = d,
                 V = V)

## Some values are so small, numerical problems.
y <- y_o
y[which(y < exp(-500), arr.ind = TRUE)] <- exp(-500)
y[which(y == 1, arr.ind = TRUE)] <- 1- exp(-500)
pos <- which(!is.na(y[,1]))
y <- y[pos,]

## Removing some rows
n <- dim(y)[1]
V <- V[pos,]

### --- 2. Precision matrix prior for latent variables --- ###
prec_prior <- 0.00001
Qx <- Matrix(diag(length(x))*prec_prior)

### --- 3. Call the function --- ####
model1 <- check1 <- dirinlareg( formula  = y ~ v1 + v2 | v1 + v2  | v1 + v2,
                               y        = y,
                               data.cov = V,
                               Qx       = Qx)


}
\author{
Joaquín Martínez-Minaya <\email{joaquin.martinez-minaya@uv.es}>
}
