% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirinlaregmodel.R
\name{dirinlaregmodel}
\alias{dirinlaregmodel}
\title{Defining a new class}
\usage{
dirinlaregmodel(
  call = NULL,
  summary_fixed = NULL,
  marginals_fixed = NULL,
  summary_linear_predictor = NULL,
  marginals_linear_predictor = NULL,
  summary_alphas = NULL,
  marginals_alphas = NULL,
  summary_precision = NULL,
  marginals_precision = NULL,
  summary_means = NULL,
  marginals_means = NULL,
  summary_predictive_alphas = NULL,
  marginals_predictive_alphas = NULL,
  summary_predictive_means = NULL,
  marginals_predictive_means = NULL,
  summary_predictive_precision = NULL,
  marginals_predictive_precision = NULL,
  dic = NULL,
  waic = NULL,
  cpo = NULL,
  nobs = NULL,
  ncat = NULL,
  y = NULL,
  data.cov = NULL
)
}
\arguments{
\item{call}{The call of the function dirinlareg.}

\item{summary_fixed}{List containing a summary of the marginal posterior distributions of the fixed effects.}

\item{marginals_fixed}{List containing the marginal posterior distributions of the fixed effects.}

\item{summary_linear_predictor}{List containing a summary of the marginal posterior distributions of the linear predictor.}

\item{marginals_linear_predictor}{List containing the marginal posterior distributions of the linear predictor.}

\item{summary_alphas}{List containing a summary of the marginal posterior distributions of the alphas.}

\item{marginals_alphas}{List containing the marginal posterior distributions of the alphas.}

\item{summary_precision}{List containing a summary of the marginal posterior distributions of the precision.}

\item{marginals_precision}{List containing the marginal posterior distributions of the precision.}

\item{summary_means}{List containing a summary of the marginal posterior distributions of the means.}

\item{marginals_means}{List containing the marginal posterior distributions of the means.}

\item{summary_predictive_alphas}{List containing a summary  of the marginal posterior predictive distribution of the alphas.}

\item{marginals_predictive_alphas}{List containing the marginal posterior predictive distribution of the alphas.}

\item{summary_predictive_means}{List containing a summary fo the marginal posterior predictive distribution of the means.}

\item{marginals_predictive_means}{List containing the marginal posterior predictive distribution of the means.}

\item{summary_predictive_precision}{List containing a summary of the marginal posterior predictive distribution of the precision.}

\item{marginals_predictive_precision}{List containing the marginal posterior predictive distribution of the precision.}

\item{dic}{List containing the inla output for dic.}

\item{waic}{List containing the inla output for waic.}

\item{cpo}{List containing the inla output for cpo.}

\item{nobs}{Number of observations.}

\item{ncat}{Number of categories.}

\item{y}{matrix containing the response variable R^{nxd}, being n number of individuals
and d the number of categories}

\item{data.cov}{data.frame with the covarites, only the covariates!}
}
\value{
object of list and dirinlaregmodel class.
}
\description{
`dirinlaregmodel` is a new object class
}
