% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fixed.R
\name{extract_fixed}
\alias{extract_fixed}
\title{Extracting marginals fixed of an inla object}
\usage{
extract_fixed(inla_model, names_cat)
}
\arguments{
\item{inla_model:}{Object of inla class.}

\item{names_cat:}{List generated with extract_formula.}
}
\value{
summary_fixed: Summary of fixed effects for each category.

marginals_fixed: Marginals for each parameter estimated.
}
\description{
`extract_fixed` is a function to extract summary and marginals distribution corresponding to the fixed effects
}
\author{
Joaquín Martínez-Minaya <\email{joaquin.martinez-minaya@uv.es}>
}
