% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_linear_predictor.R
\name{extract_linear_predictor}
\alias{extract_linear_predictor}
\title{Extracting posterior distributions from the linear predictor}
\usage{
extract_linear_predictor(inla_model, n, d, Lk_eta, names_cat = names_cat, sim)
}
\arguments{
\item{inla_model}{An object of class inla.}

\item{n}{Number of observations.}

\item{d}{Number of categories.}

\item{Lk_eta}{Cholesky decomposition of the Hessian matrix.}

\item{names_cat}{List generated with extract_formula.}

\item{sim}{simulations for the function inla.posterior.sample}
}
\value{
summary_linear_predictor List containing a summary of the marginal posterior distributions of the linear predictor.

marginals_linear_predictor List containing simulations of marginal posterior distributions of the linear predictor.

summary_alphas List containing a summary of the marginal posterior distributions of the alphas.

marginals_alphas List containing simulations of the marginal posterior distributions of the alphas.

summary_precision List containing a summary of the marginal posterior distributions of the precision.

marginals_precision List containing simulations of the marginal posterior distributions of the precision.

summary_means List containing a summary of the marginal posterior distributions of the means.

marginals_means List containing the simulations of the marginal posterior distributions of the means.
}
\description{
`extract_linear_predictor` extracts the posterior distribution from the linear predictor
}
\author{
Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>
}
