% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_list.R
\name{formula_list}
\alias{formula_list}
\title{Formula in to list}
\usage{
formula_list(form)
}
\arguments{
\item{form}{Object of class formula.}
}
\value{
A list with the names of the variables used in each category.
}
\description{
`formula_list` reads the formula and generates a list with the name of the covariates used in each category
}
\examples{
formula <- y ~ 1 + v1 + v2 | -1 + v1 | 0 + v2
formula_list(formula)

}
\author{
Joaquín Martínez-Minaya <\email{joaquin.martinez-minaya@uv.es}>
}
