% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H0_matrix_eta1.R, R/H0_matrix_eta_x.R
\name{H0_matrix_eta1}
\alias{H0_matrix_eta1}
\title{Computing expected Hessian for a vector eta}
\usage{
H0_matrix_eta1(eta, d)

H0_matrix_eta1(eta, d)
}
\arguments{
\item{eta}{eta vector to compute the expected Hessian.}

\item{d}{Dimension}
}
\value{
Expected Hessian in eta.

Expected Hessian in eta.
}
\description{
`H0_matrix_eta_1` computes the expected Hessian in eta of -loglikelihood

`H0_matrix_eta_1` computes the expected Hessian in eta of -loglikelihood
}
\author{
Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>

Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>
}
