% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H0_matrix_eta1.R, R/H0_matrix_eta_x.R
\name{H0_matrix_eta_x}
\alias{H0_matrix_eta_x}
\title{Computing expected Hessian in eta}
\usage{
H0_matrix_eta_x(eta, d, cores)

H0_matrix_eta_x(eta, d, cores)
}
\arguments{
\item{eta}{Linear predictor resulting of the product Ax.}

\item{d}{Dimension.}

\item{cores}{Number of cores for parallel computation. The package parallel is used.}
}
\value{
Expected Hessian in eta.

Expected Hessian in eta.
}
\description{
`H0_matrix_eta_x` computes the expected Hessian in eta of -loglikelihood

`H0_matrix_eta_x` computes the expected Hessian in eta of -loglikelihood
}
\author{
Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>

Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>
}
