% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_beta_mult_eta.R
\name{log_beta_mult_eta}
\alias{log_beta_mult_eta}
\title{Calculating the log beta function in eta}
\usage{
log_beta_mult_eta(x)
}
\arguments{
\item{x}{Vector of elements.}
}
\value{
Numeric value.
}
\description{
\code{beta_mult_eta} computes the log beta function in eta
}
\author{
Joaquín Martínez-Minaya \email{joaquin.martinez-minaya@uv.es}
}
