% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton_x.R
\name{newton_x}
\alias{newton_x}
\title{Newton-Raphson algorithm}
\usage{
newton_x(A, x_hat, gk, Hk, a, Qx, strategy, y, d = d)
}
\arguments{
\item{A:}{Matrix which links eta with the latent field, i.e., eta = A x}

\item{x_hat:}{Vector with the elements of the latent field, i.e., eta_hat = A x_hat}

\item{gk:}{Gradient in eta.}

\item{Hk:}{Hessian in eta.}

\item{a:}{Step length.}

\item{Qx:}{Precision matrix for the prior of the latent field.}

\item{strategy:}{Strategy to use to optimize. Now, line search strategy with quasi-newton algorithm is the only one avaliable.}

\item{y:}{Vector with the response variable}

\item{d:}{Number of categories.}
}
\value{
g0 : Gradient in x_hat_new. A numeric vector with the gradient in x_hat_new.

x_hat_new: New value of x after apply one iteration.
}
\description{
`newton_x` computes optimization algorithms to find the mode of the
posterior. Line search strategy with Armijo conditions is implemented
}
\author{
Joaquín Martínez-Minaya <\email{joaquin.martinez-minaya@uv.es}>
}
