% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dirinlaregmodel.R
\name{plot.dirinlaregmodel}
\alias{plot.dirinlaregmodel}
\title{plot of dirinlaregmodel objects}
\usage{
plot.dirinlaregmodel(object)
}
\arguments{
\item{object}{Object of class dirinlareg.}
}
\value{
Plotting the posterior of the fixed effects.
}
\description{
`summary.dirinlaregmodel` Method which plots a dirinlaregmodel object
}
\author{
Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>
}
