% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dirinlaregmodel.R
\name{plot.dirinlaregmodel}
\alias{plot.dirinlaregmodel}
\title{plot of dirinlaregmodel xs}
\usage{
plot.dirinlaregmodel(x, ...)
}
\arguments{
\item{x}{Object of class dirinlaregmodel.}

\item{...}{Other arguments.}
}
\value{
Plotting the posterior of the fixed effects.
}
\description{
`plot.dirinlaregmodel` Method which plots a dirinlaregmodel x
}
\author{
Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>
}
