% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dirinlaregmodel.R
\name{summary.dirinlaregmodel}
\alias{summary.dirinlaregmodel}
\title{Summary of dirinlaregmodel objects}
\usage{
\method{summary}{dirinlaregmodel}(object, ...)
}
\arguments{
\item{object}{Object of class dirinlaregmodel.}

\item{...}{Other arguments.}
}
\value{
Print summary.
}
\description{
`summary.dirinlaregmodel` is a function which gives a summary of a dirinlaregmodel object
}
\author{
Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>
}
