% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fast.R
\name{summary_fast}
\alias{summary_fast}
\title{Summary using the packages Rfast and Rfast2 of a matrix by rows}
\usage{
summary_fast(A)
}
\arguments{
\item{A}{matrix to be summarised}
}
\value{
A matrix whose columns are "mean", "sd", "0.025quant", "0.5quant", "0.975quant"
}
\description{
`summary_fast` summarise a matrix by rows
}
\examples{
A <- matrix(rnorm(10000), ncol = 1000)
summary_fast(A)
}
\author{
Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>
}
