% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fast.R
\name{summary_fast}
\alias{summary_fast}
\title{Summary using the packages Rfast and Rfast2 of a matrix by rows}
\usage{
summary_fast(A)
}
\arguments{
\item{A}{matrix to be summarised}
}
\value{
A matrix whose columns are "mean", "min", "q0.025", "q0.25", "q0.5", "q0.75", "q0.975", "max"
}
\description{
`summary_fast` summarise a matrix by rows
}
\author{
Joaquín Martínez-Minaya <\email{jomarminaya@gmail.com}>
}
