% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digamma.R
\name{trigamma_red}
\alias{trigamma_red}
\title{Computing the function trigamma}
\usage{
trigamma_red(x, ...)
}
\arguments{
\item{x}{Argument to applied the function trigamma.}

\item{...}{Rest of arguments used in the case of digamma functions.}
}
\value{
Result of applying trigamma function.
}
\description{
\code{trigamma_red} is the function trigamma appropiate for really small values
}
\author{
Joaquín Martínez-Minaya \email{joaquin.martinez-minaya@uv.es}
}
